+++ dflib-echarts/src/main/java/org/dflib/echarts/LineSymbol.java
@@ -3,6 +3,7 @@ package org.dflib.echarts;
 /**
  * @since 2.0.0
  */
+// TODO: same symbols seem to be used by scatter plots. So the name "line" is misleading
 public enum LineSymbol {
     emptyCircle, circle, rect, roundRect, triangle, diamond, pin, arrow, none;
 }